decouple vm context and validation to two files
