+++ src/views/catalog/utils/WizardVMContext/utils/vm-produce.ts
